Calendar UI update
added some drawables to navigation drawer.

replaced settings menu in app bar with a ‘TODAY’ button (moved options
in settings menu to nav drawer).

changed orientation layout design to happen on large devices (tablets)
only (to satisfy spec)

removed some old classes and xml of mine that are redundant.

added functionality to retain the currently viewed date (and time) when
changing orientation as well as the number of days in calendar UI.