Added working Calendar Provider
a List<Event> of all events can now be called. It may need filtering at
some point and not sure if timezones need to be messed around with to
make the times accurate but we shall see when it gets used with the
calendar views.

See the Main Activity for an example on how to output events.